requirements.txt
lameenc
audioread
yt-dlp
kivy
loguru
beautifulsoup4
requests
eyeD3
Pillow
kivymd2